+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2.3.1 (...)
+## 2.3.1 (2016-02-20)
 - static IP can also not be network and broadcast of `poolRange`
 - do not allow to set static IP addresses already in use
 - make IP ranges available through API for edit page